test: extract getting items into helper class (#7726)
