+++ vaadin-virtual-list-flow-parent/vaadin-virtual-list-flow-integration-tests/src/test/java/com/vaadin/flow/component/virtuallist/tests/VirtualListHelpers.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2000-2025 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.vaadin.flow.component.virtuallist.tests;
+
+import java.util.List;
+import java.util.Map;
+
+import org.openqa.selenium.JavascriptExecutor;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+
+import elemental.json.Json;
+import elemental.json.JsonArray;
+import elemental.json.JsonObject;
+
+public class VirtualListHelpers {
+
+    public static JsonArray getItems(WebDriver driver, WebElement element) {
+        Object result = ((JavascriptExecutor) driver)
+                .executeScript("return arguments[0].items;", element);
+        JsonArray array = Json.createArray();
+        if (!(result instanceof List)) {
+            return array;
+        }
+        List<Map<String, ?>> list = (List<Map<String, ?>>) result;
+        for (int i = 0; i < list.size(); i++) {
+            Map<String, ?> map = list.get(i);
+            if (map != null) {
+                JsonObject obj = Json.createObject();
+                map.entrySet().forEach(entry -> {
+                    obj.put(entry.getKey(), String.valueOf(entry.getValue()));
+                });
+                array.set(i, obj);
+            } else {
+                array.set(i, Json.createNull());
+            }
+        }
+        return array;
+    }
+}
